backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlmodel==0.0.14
psycopg2-binary==2.9.9
PyJWT==2.8.0
pwdlib[argon2]==0.2.1
python-multipart==0.0.6
pydantic[email]==2.5.3
alembic==1.13.1
python-dateutil==2.8.2
